move accordion/accordion.gts to accordion.gts
